Correctly configure buttons for complications.
This commit excludes the background complication and any fixed
complication from the list of complications that can be clicked in the
configuration screen of the editor sample. This is because the
background complication can be configured by means of its own dedicate
entry in the settings activity and fixed complications cannot be
configured at all.

This fixes a problem with the sample that, since the background
complication was covering the entire screen and was above all the
other complication buttons, it was not possible to configure any
complication besides the background one.

Bug: 186353928
Test: ./gradlew :wear:wear-watchface-samples-app:installDebug and manually validate the complications
Change-Id: I5b2e815b17da2e42470325328cfd293930f18d09