Enable slots reuse in LazyColumn
Lazy lists now set 2 as maxSlotsToRetainForReuse param for SubcomposeLayoutState. This improves the performance of composing next item as we reuse active compositions and LayoutNode objects. This especially helps if all the items have the similar layout hierarchy.
This also changes the current behavior where previously the item which was scrolled out was immediately disposed. Now this item can be kept active til reused.

Relnote: LazyColumn/Row will now keep up to 2 previously visible items active (not disposed) even when they are scrolled out already. This allows the component to reuse the active subcompositions when we will need to compose a new item which improves the scrolling performance.
Test: LazyListSlotsReuseTest
Change-Id: Ie5555c9a7031dc9bd31f452a0ed9b28d8a337f5f