Add the root of sub composition to content
A Dialog, AlertDialog, and Popup will appear fragmented in the
SlotTable with some Group nodes in the app and other Group nodes
in a separate window.

This change will stitch these parts together such that the user
will be able to see e.g. the "AlertDialog" node as an immediate
parent to the content of the dialog.

Bug: 180648394
Test: Unit tests added
Change-Id: I88bc233f0ca9f072540e8ebde943fc83849937a8