Fix GeneratedCodeMatchTest for KoltinPoet 1.8
In KotlinPoet 1.7 explicit API mode is turned ON by default and this has an effect in the code generated, specifically visibility keywords are emitted. Since the test involves a mismatch in expected vs generated sources, the change in API mode affected the test and its outcome. This CL simply updates the expected and actual sources with the visibility keyword to align with the KotlinPoet update.

Test: GeneratedCodeMatchTest
Change-Id: I8e2fa53f051a914125b95cdcffe4469532946890