Added getters for all builders.
This also includes wrapper classes for a bunch of previously
hidden messages. As an example, any builder that takes a String, would
typically have internally translated this to a StringProp message, which
we use to allow us to add additional metadata in future if needed.
Because of this, we can't add methods for `string getFoo()`, as this
would prevent us from also returning the metadata in future.

Instead, we expose these previously hidden messages. This means that:

    public Builder setFoo(string foo)

Now generates the following:

    // Old shortcut
    public Builder setFoo(string foo)

    // Version that takes the previously hidden message
    public Builder setFoo(StringProp foo)

    // New getter
    public StringProp getFoo()

Of course, to support this, we also had to add the class definitions for
StringProp etc that were previously implementation details.

Bug: 185912965
Test: Existing tests work; manually tested.
Relnote: Added getters for all builders, to enable easier testing.
Change-Id: I9d1554305b55d2e3c2676084fe8a6f0ef5e51aeb