Fix layout bug when placing Images in a Box.
This fixes two bugs:

* We attempted to ignore intrinsic images sizes, but in certain
conditions, they could be still be used (e.g. when adding a large image
to a Box with size=wrap(), the Box could still use the image intrinsic
size).
* Due to a bug in FrameLayout, a Box that contains a single expand()
child won't force the dimensions of the child.

Bug: 187512483
Relnote: Fix layout bug when placing Images in a Box.
Test: New golden tests incoming
Change-Id: I33770e987e419a0d03ae95e7ddceb42904ace9f7