Add AnimatedVisibility tooling support
Instead of ignoring AnimatedVisibility calls, detect them and track the
animations a the newly created ComposeAnimationType.ANIMATED_VISIBILITY.

AnimatedVisibilityComposeAnimation's overridden `animationObject` API points
to the AnimatedVisibility's parent (Boolean) transition, and we added a new
property (`childTransition`) to point to its single child (EnterExitState).
The parent is used when seeking, while the child is used when retrieving the
animated properties.

Handling AnimatedVisibility as a regular transition would allow developers
to set invalid state combinations, e.g. true -> true if controlling the
parent or PreEnter -> PostExit if controlling the child. We work around that
by creating an intermediate state (AnimatedVisibilityState), which can be
either Enter or Exit.

Bug: 187962923
Bug: 178910730
Test: PreviewAnimationClockTest and ComposeViewAdapterTest
Relnote: N/A. API changes related to animation-tooling-internal only.
Change-Id: I6a6ecf02c3c140360925487f83177c6df2e81ddd