Hide ActionBar Up Icon on top level destinations
When using the setupActionBarWithNavController()
method, just setting the up indicator to null
isn't enough to remove the Up icon - we need
to specifically call setDisplayHomeAsUpEnabled(false).

This fixes a regression when this class
was moved to Kotlin.

As part of this, moved to checkNotNull() around
nullable properties to better match the
expectations that these should never be null.

Test: tested in NavigationAdvancedSample
BUG: 189868637
Relnote: "Fixed a regression introduced in
[Navigation `2.4.0-alpha01`](#2.4.0-alpha01) where
`setupActionBarWithNavController()` would not
properly remove the Up icon when you are on a
top level destination."

Change-Id: I5d4439021161f544d664a822e1c422ff60c91e1b