Introduce AnimatedContent
This CL introduces a new composable AnimatedContent. It
is responisble for:
1) Managing the enter/exit transition of children
2) Automatically animating the container size to the target content
3) Supporting enter/exit/size animation/z-order to be customized
   via ContentTransform.

This composable is designed to be a versatile layer that can be used
in a variety of use cases such as screen transition, and any
 state-dependent content. All animations created in AnimatedContent
are put on a Transition object. This allows tooling to later
support visualization of the content transform.

Relnote: "New `AnimatedContent` Composable. It manages its content
change using customizable `ContentTransform` as new target content
enters and initial content leaves. Different combination of enter
and exit transitions can be used to produce a customized look and
feel.  As a part of the content transform, `AnimatedContent`
automatically animates its size to match the incoming content. "

Test: Included

Change-Id: I2c3df9070957cc144b13b404848b7fe4717dd2b8