Fix conflicts between remove by pop and hide
If you do an add, then call popBackStack() and commit a transaction to
hide the same fragment is being removed by the pop while using
transitions and not using `reorderingAllowed(true)`, the fragment
manager will end up destroying the view of the fragment too early. The
reason for this is that the pop will start the transition of the remove,
but before the remove is complete, the hide operation is enqueued and
causes the computed state of the fragment to be incorrect and that
results in the view being destroyed.

While hiding a removed fragment is not something you actually probably
want to do, we should make sure this at least doesn't crash.

Instead of relying solely on the new pending operations to determine the
state of the fragment, if the pending operation does not affect the
Lifecycle (i.e. it is a show or hide), we should also consider the
lifecycle impact of the running operation as well.

RelNote: "The `FragmentManager` will no longer crash when you attempt to
hide a removing fragment."
Test: testPopRemoveWithHide()
Bug: 183634730

Change-Id: I573ddcfcd1b9c782b5546dd4643e0e4ed042b277