Get applicationId using reflection
Using safe args with dynamic feature modules with AGP 7.0+ currently
crashes saying that DynamicFeatureVariant does not have the
applicationId property. It does have the property, but in 7.0+ it is
part of the GeneratesApk interface that DynamicFeatureVariant
implements and the build cannot resolve that DynamicFeatureVariant has
that method.

We need to use reflection to call the property dynamically instead.

RelNote: "Safe-args no longer crashes when using it with dynamic feature
modules and AGP 7.0+."
Test: tested with sample apps
Bug: 189966576

Change-Id: I6951812da260e739a2d515fb9b3782b90c75642d