Remove uses of titleCase in Safe Args
The titleCase API was added in Kotlin 1.5.0. Gradle versions previous to
7.0 relied on eariler versionss of Kotlin that do not have that API
available. SafeArgs now depends on Android Gradle Plugin 4.2.0, which is
compatible down to Gradle 6.7.1, so we should continue to support older
versions of kotlin until we move to the next stable version of AGP.

RelNote: "SafeArgs will no longer fail when building your app with
`Kotlin` versions before `1.5.0`."
Test: tested in sample app
Bug: 190739257

Change-Id: Icd1ff3c64f220f4b310934c2edf910d6aae01475