Don't peek Wearable drawers on horizontal views.
WearableDrawerLayout attaches fling listeners onto the views inside it.
These are intended to detect the end of a fling, then if the view has
reached the top/bottom, they will cause the drawers to peek. If the
WearableDrawerLayout contains a horizontal RecyclerView though (which
can happen via ViewPager2), the drawers would peek on any scroll, since
the top/bottom detection is done by just checking whether the View could
be vertically scrolled.

The CL detects this case by checking if the inner view can be vertically
scrolled at all. If not, then the peeking is inhibited.

Relnote: "Fixed bug where horizontally scrolling RecyclerViews caused
WearableDrawerLayout to peek on all interactions."
Test: Extended WearableDrawerLayoutEspressoTest

Change-Id: I24c7f3a4afecdad53b2952df11e6dfdab82050d3