Add CurvedRow to wear-compose
A CurvedRow is a Layout that places its children on an arc, similar to
a row, but with the content drawn in an annulus sector instead of a
box.
Parameters can be used to specify which angle will be used as anchor to
lay down the children, in which direction to laid them down (clockwise
or counter-clockwise), and how to use the anchor (start there, end there
or center around there).
Added a new category for foundation demos and a small screen to the Demo.
Added a new project for samples to be used in the demo and javadocs.

Test: To be added.
Relnote: "Add a new class CurvedRow to laid out composables in an arc"
Bug: 190807301

Change-Id: I29941d2f915c3da65194402367dcd30898a4d0b2