Revert "Make scrollable to adjust velocity if some delta consumed during "pre"-stage of the nested scroll"
This reverts commit 0d28e2eb472172f385fd7c89c16e4f5d69cfe591.

Reason for revert: It's a harmful fix, rendering nested scroll unusable for cases when the parent consumes all the delta all the time. This change also breaks the geometry of velocity as a whole, so the velocity is not following the finger physics.

Relnote: n/a
Test: n/a
Change-Id: I23de94bba5e58d1c04cfed2b653788005d6bf968
(cherry picked from commit 2f4b994183ae01328eb67521186c4b7272a01e37)