Avoid processing emoji after InputFilter removed
EmojiInputFilter previously would process emoji for all text views that
were disabled between instantiation and EmojiCompat finishing loading
the font.

This change makes EmojiInputFilter check to ensure that it's still
registered as an InputFilter on the TextView prior to processing emoji
in response to font load callbacks.

This is purely a bugfix, and will avoid situations where TextViews with
emoji processing disabled would process emoji once on font load
completion. This bug was discovered when researching theme-level
emojiCompatEnabled=false.

Bug: b/188083840
Test: New unit test
Change-Id: I4569227df84944de65cbb20274be8f5eaaaaa66e