Rename param in coordinatorlayout, core, and wear libraries
This is a continuation of aosp/1751886, with modifications to the coordinatorlayout and wear libraries as part of the process for integrating the core library with Dackka.

If method A has JavaDoc for its param name and method B overrides method A, method B's param variable(s) must match the spelling of method A's param variable(s).  If they don't match, Dackka errors when processing the JavaDoc.

This also fixes a typo in a <ol> closing tag.

Bug: 19208685
Test: presubmit
Change-Id: I6da2b524178743ce3b5783f0be9105848f101a11