Simplify synthetic getter/setter methods
In the early days of KSP, it didn't create getter/setter accessors for
properties unless it was explicitly declared so we worked around it by
creating a synthetic one.

Now it does create it always, which means we can simplify our logic.
Note that these are still not KSFunctionDeclaration implementations so
we cannot simply use the declared function infra so we still have the
synthetics, just much simpler.

Bug: n/a
Test: existing tests
Change-Id: I8e059c1c6ca9c3a85471e0d87d8951e2567186fe