[RevokePermissions] Add binding to the UnusedAppRestrictionsBackportService in order to retrieve the permission revocation setting and return it in getUnusedAppRestrictionsStatus.
This change builds on top of aosp/1757273. It adds binding to the
UnusedAppRestrictionsBackportService in order to fetch the permission
revocation setting from the backport and does so in the separate
PackageManagerCompat class.

BYPASS_INCLUSIVE_LANGUAGE_REASON=Using an existing Android R API
(isAutoRevokeWhitelisted)
Relnote: Completed functionality for getUnusedAppRestrictionsStatus.
BUG: 191919561
FIXED: 177234481
FIXED: 192829937
Test: IntentCompatTest, PackageManagerCompatTest,
UnusedAppRestrictionsBackportServiceConnectionTest

Change-Id: I7c2d63bb73588c8db98fdde4b13be652df699b4a