Properly handle Object types in override variance
This CL fixes a bug where if Unit is used in type arguments, we were
creating `out Unit` even when we shouldn't.

Seems like the check was including OBJECT classkinds by default. I'm not
sure why, couldn't find it in CL notes but that seems unnecessary. Added
a test to validate that it is not necessary (also, Unit is an `object`)

Added a couple more tests with parmeters that have function types.

Fixes: 190010217
Test: MethodSpecHelperTest
Change-Id: Ia628235a7385d3561ed7fb0eacfa9ebe486b75d5