Handle type aliases in getAnnotation
This CL fixes a bug in XProcessing where we were not resolving
type aliases while checking annotation types, resulted in missing
return types when the source uses type aliases.

Unfortunately, KSP also has a bug that affects getSymbolsWithAnnotation
but at least we can fix one part of it in XProcessing.
https://github.com/google/ksp/issues/506

Bug: 188853151
Test: XAnnotationTest
Change-Id: Iaf2501d6c27c9c30113a044b33cf4efd27746b18