Reduce throttle max retries, and reset if hitting max retries
Fixes: 193260686
Test: ./gradlew benchmark:benchmark-junit4:cC benchmark:benchmark-common:cC
Relnote: """Reduce aggressiveness of thermal throttle detection, and
recompute baseline if throttles are detected frequently."""

Reduce max throttle checks in a row, to reduce cost of incorrect
throttle detection. In CI, we're seeing BenchmarkRuleTest hit thermal
throttling several times in a row on low API devices, possibly because
the first measurement is at high launch clocks, and others are not.

Because 3 throttle backoffs in a row (270+ seconds) essentially
guarantee we hit the @LargeTest timeout (300 seconds), reducing this
should significantly.

Additionally, at the end of a benchmark, reset throttle baseline if we
ran out of throttle retries, as it may indicate an incorrect initial
baseline like the situations in CI. This portion is a somewhat
speculative fix, but the CI logs show we've been timing out in this
test several times, and this prevents incorrect state from being
stored and used throughout an entire test suite run.

Change-Id: I7327b4f1c15317fad28de1f06a148aad29c7cd4f