Remove androidx.annotation dependency from room-compiler-processing.
The main usage of the dependency was to find nullability annotations using the class, but instead we can find those using the class's canonical name.

The second usage of the dep was for @VisibleForTesting, one usage was easily removed (KspReflectiveAnnotationBox) and the other (in JavacTestProcessor) required a bit of refactor.

Fixes: 192608164
Test: ./gradlew :room:r-c-p:test
Change-Id: I5d95fa328f0d875dbb3b1d8a562a4028c518cbfe